+++ src/vector/feature.rs
@@ -47,6 +47,11 @@ impl<'a> Feature<'a> {
         }
     }
 
+    /// Returns the C wrapped pointer
+    pub fn c_feature(&self) -> OGRFeatureH {
+        self.c_feature
+    }
+
     pub fn _lazy_feature_geometries(defn: &'a Defn) -> Vec<Geometry> {
         let geom_field_count =
             unsafe { gdal_sys::OGR_FD_GetGeomFieldCount(defn.c_defn()) } as isize;