+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add `Feature::c_feature` that returns the OGR feature handle.
+  - <https://github.com/georust/gdal/pull/192>
 - Add wrapper for `OGR_G_Buffer`.
 - Add support for raster dataset creation options. A new struct (`RasterCreationOption`) and function (`driver.create_with_band_type_with_options()`) are now available for this.